# exercise-execution-service/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Exercise Execution Service API
  description: |
    API REST para registro de execuções de exercícios físicos.

    ## Autenticação
    Rotas protegidas exigem um token JWT no header `Authorization: Bearer <token>`.
    Obtenha o token via `POST /auth/login`.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
          example: 1
        slug:
          type: string
          example: pushup
        name:
          type: string
          example: Flexão
        description:
          type: string
          example: Apoie as mãos no chão na largura dos ombros...
        createdAt:
          type: string
          format: date-time
    Execution:
      type: object
      properties:
        id:
          type: integer
          example: 11
        userId:
          type: integer
          example: 1
        exerciseId:
          type: integer
          example: 1
        reps:
          type: integer
          example: 15
        durationSec:
          type: integer
          example: 45
        result:
          type: string
          enum: [correct, incorrect]
          example: correct
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.92
        executedAt:
          type: string
          format: date-time
          example: '2026-04-21T12:00:00.000Z'
        createdAt:
          type: string
          format: date-time
        exercise:
          $ref: '#/components/schemas/Exercise'
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: object
paths:
...
  /executions/{id}:
    get:
      summary: Detalha execução do usuário autenticado
      tags: [Executions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Execução encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution:
                    $ref: '#/components/schemas/Execution'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execução não encontrada (ou pertence a outro usuário)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Atualiza execução do usuário autenticado
      tags: [Executions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reps:
                  type: integer
                  example: 20
                durationSec:
                  type: integer
                  example: 60
                result:
                  type: string
                  enum: [correct, incorrect]
                score:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 1
                  example: 0.95
                executedAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Execução atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution:
                    $ref: '#/components/schemas/Execution'
        '400':
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execução não encontrada (ou pertence a outro usuário)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove execução do usuário autenticado
      tags: [Executions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: Execução removida (sem corpo)
        '401':
          description: Token ausente ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Execução não encontrada (ou pertence a outro usuário)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'